backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
pillow==10.1.0
PyGithub==2.1.1
python-social-auth==0.3.6
moviepy==1.0.3
pytest==7.4.3
alembic==1.12.1
psycopg2-binary==2.9.9 